+++ previous-versions/commerce/src/sdk/sdk.ts
@@ -62,9 +62,9 @@ export class SDKConfiguration {
     serverDefaults: any;
     language = "typescript";
     openapiDocVersion = "3.0.0";
-    sdkVersion = "0.37.0";
-    genVersion = "2.150.0";
-    userAgent = "speakeasy-sdk/typescript 0.37.0 2.150.0 3.0.0 @codat/commerce";
+    sdkVersion = "0.37.1";
+    genVersion = "2.155.1";
+    userAgent = "speakeasy-sdk/typescript 0.37.1 2.155.1 3.0.0 @codat/commerce";
     retryConfig?: utils.RetryConfig;
     public constructor(init?: Partial<SDKConfiguration>) {
         Object.assign(this, init);